# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the WhatsApp newspaper kiosk bot.",
        "title": "WhatsApp Bot API",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/file": {
            "post": {
                "description": "Notifies all subscribers whose subscription text matches the provided filename.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiosk"
                ],
                "summary": "Trigger subscription notifications",
                "parameters": [
                    {
                        "description": "Subscription trigger",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.triggerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.statusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.statusResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/health": {
            "get": {
                "description": "Returns the current health status of the application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HealthInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications": {
            "post": {
                "description": "Sends a plain text message to the given WhatsApp recipient.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Whatsapp"
                ],
                "summary": "Send a WhatsApp notification",
                "parameters": [
                    {
                        "description": "Notification",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.notificationRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/handlers.notificationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/version": {
            "get": {
                "description": "Returns build and version metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get version information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VersionInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/whatsapp/groups": {
            "get": {
                "description": "Returns the WhatsApp groups the bot is participating in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Whatsapp"
                ],
                "summary": "List WhatsApp groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.groupResponse"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/whatsapp/groups/{jid}/participants": {
            "get": {
                "description": "Returns the participants of the given WhatsApp group.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Whatsapp"
                ],
                "summary": "List group participants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group JID",
                        "name": "jid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.participantResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/internacional_folder/{file}": {
            "get": {
                "description": "Downloads a file from the configured kiosk folder. Requires a valid JWT token passed as the access_token query parameter.",
                "tags": [
                    "Quiosk"
                ],
                "summary": "Download a kiosk file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "file",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "access_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/magazine_folder/{file}": {
            "get": {
                "description": "Downloads a file from the configured kiosk folder. Requires a valid JWT token passed as the access_token query parameter.",
                "tags": [
                    "Quiosk"
                ],
                "summary": "Download a kiosk file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "file",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "access_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Exposes Prometheus-compatible metrics.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Prometheus metrics",
                "responses": {
                    "200": {
                        "description": "Prometheus metrics text",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/nacional_folder/{file}": {
            "get": {
                "description": "Downloads a file from the configured kiosk folder. Requires a valid JWT token passed as the access_token query parameter.",
                "tags": [
                    "Quiosk"
                ],
                "summary": "Download a kiosk file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "file",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "access_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.groupResponse": {
            "type": "object",
            "properties": {
                "jid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.notificationRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                }
            }
        },
        "handlers.notificationResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.participantResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "is_super_admin": {
                    "type": "boolean"
                },
                "jid": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "handlers.statusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.triggerRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "file": {
                    "type": "string"
                }
            }
        },
        "model.HealthInfo": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.VersionInfo": {
            "type": "object",
            "properties": {
                "appName": {
                    "type": "string"
                },
                "buildTime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}